0.9.4 initial, issues with UI hooks on placement
