+++ JotunnModStub/Challenge/PortalTracker.cs
@@ -10,6 +10,7 @@ namespace ValheimFortress.Challenge
     internal class PortalTracker : MonoBehaviour
     {
         private GameObject shrineReference;
+        private static int destroy_timer = 0;
         public void SetShrine(GameObject shrine)
         {
             shrineReference = shrine;
@@ -17,6 +18,16 @@ namespace ValheimFortress.Challenge
 
         public void Update()
         {
+            // if the shrine reference is not set we start a countdown timer to destroy the portal
+            // This only occurs when the portal is orphaned
+            if (shrineReference == null) {
+                if (destroy_timer >= 120) { Destroy(gameObject); }
+
+                destroy_timer++;
+                return;
+            }
+
+            // If the shrine reference is set we check to see if a challenge is active and when its no longer active we destroy the portal
             if (shrineReference.GetComponent<Shrine>().challenge_active.Get() == false)
             {
                 GameObject destroyVFX = UnityEngine.Object.Instantiate(ValheimFortress.getPortalDestroyVFX(), this.transform.position, this.transform.rotation);
