+++ JotunnModStub/Challenge/Spawner.cs
@@ -159,6 +159,8 @@ namespace ValheimFortress.Challenge
                 Quaternion rotation = Quaternion.Euler(0f, UnityEngine.Random.Range(0f, 360f), 0f);
                 var tempportal = UnityEngine.Object.Instantiate(ValheimFortress.getPortal(), spawn_location, rotation);
                 portals.Add(tempportal);
+                // We are going to manipulate the lifetime of this object outside of its timed life.
+                Destroy(tempportal.GetComponent<TimedDestruction>());
                 tempportal.AddComponent<PortalTracker>();
                 tempportal.GetComponent<PortalTracker>().SetShrine(shrine);
                 if (VFConfig.EnableMapPings.Value) { Chat.instance.SendPing(spawn_location); }
