+++ JotunnModStub/Challenge/UI.cs
@@ -9,65 +9,51 @@ namespace ValheimFortress.Challenge
 {
     public class UI : MonoBehaviour
     {
-        static GameObject ChallengePanel;
-        static GameObject CancelPanel;
+        private GameObject ChallengePanel;
+        private GameObject CancelPanel;
         // this is unset until the shrine building calls for the UI, in which case it is then set
-        static Shrine Shrine;
+        private Shrine Shrine;
 
-        static List<String> currentLevels = new List<String> {};
-        static List<String> availableRewards = new List<String> {};
-        public static GameObject levelSelector;
-        public static GameObject rewardSelector;
-        public static GameObject hardModeToggle;
-        public static GameObject bossModeToggle;
-        public static GameObject siegeModeToggle;
-        public static GameObject estimate_text;
-        public static short estimatedRewards = 0;
-        public static string estimatedRewardName = "";
+        private List<String> currentLevels = UserInterfaceData.currentLevels;
+        private List<String> availableRewards = UserInterfaceData.availableRewards;
+        public GameObject levelSelector;
+        public GameObject rewardSelector;
+        public GameObject hardModeToggle;
+        public GameObject bossModeToggle;
+        public GameObject siegeModeToggle;
+        public GameObject estimate_text;
+        public short estimatedRewards = 0;
+        public string estimatedRewardName = "";
 
-        private static GameObject estimate_symbol;
-        private static GameObject hardmode_label;
-        private static GameObject hardmode_desc;
-        private static GameObject hardmode_reward_desc;
-        private static GameObject bossmode_label;
-        private static GameObject bossmode_desc;
-        private static GameObject bossmode_reward_desc;
-        private static GameObject siegemode_label;
-        private static GameObject siegemode_desc;
-        private static GameObject siegemode_reward_desc;
+        private GameObject estimate_symbol;
+        private GameObject hardmode_label;
+        private GameObject hardmode_desc;
+        private GameObject hardmode_reward_desc;
+        private GameObject bossmode_label;
+        private GameObject bossmode_desc;
+        private GameObject bossmode_reward_desc;
+        private GameObject siegemode_label;
+        private GameObject siegemode_desc;
+        private GameObject siegemode_reward_desc;
 
+        private GameObject cancelButtonGO;
+        private GameObject startChallengeButtonGO;
 
-        public static short selected_level = 0;
-        public static bool current_hard_mode = false;
-        public static bool current_boss_mode = false;
-        public static bool current_siege_mode = false;
+        public short selected_level = 0;
+        public bool current_hard_mode = false;
+        public bool current_boss_mode = false;
+        public bool current_siege_mode = false;
 
-        // Right now maxlevel needs to correlate to: defined levels & level warning messages
-        private static List<String> shrine_phase_warnings = new List<String>
+        public void Awake()
         {
-            Localization.instance.Localize("$shrine_phase_warning"),
-            Localization.instance.Localize("$shrine_phase_warning2"),
-            Localization.instance.Localize("$shrine_phase_warning3"),
-            Localization.instance.Localize("$shrine_phase_warning4"),
-            Localization.instance.Localize("$shrine_phase_warning5"),
-            Localization.instance.Localize("$shrine_phase_warning6"),
-            Localization.instance.Localize("$shrine_phase_warning7"),
-            Localization.instance.Localize("$shrine_phase_warning8"),
-            Localization.instance.Localize("$shrine_phase_warning9"),
-            Localization.instance.Localize("$shrine_phase_warning10"),
-            Localization.instance.Localize("$shrine_phase_warning11"),
-            Localization.instance.Localize("$shrine_phase_warning12"),
-            Localization.instance.Localize("$shrine_phase_warning13"),
-            Localization.instance.Localize("$shrine_phase_warning14"),
-            Localization.instance.Localize("$shrine_phase_warning15"),
-            Localization.instance.Localize("$shrine_phase_warning16"),
-            Localization.instance.Localize("$shrine_phase_warning17"),
-            Localization.instance.Localize("$shrine_phase_warning18"),
-            Localization.instance.Localize("$shrine_phase_warning19")
-        };
+            Shrine = this.GetComponent<Shrine>();
+        }
 
-        public static bool IsPanelVisible()
+        public bool IsPanelVisible()
         {
+            // Challenge panel doesn't exist yet, so no its not visible
+            if (ChallengePanel == null) { return false; }
+
             return ChallengePanel.activeSelf;
         }
 
@@ -110,103 +96,6 @@ namespace ValheimFortress.Challenge
             }
         }
 
-        public void Awake()
-        {
-            Shrine = this.GetComponent<Shrine>();
-            CreateStaticUIObjects(true);
-            // CreateChallengeUI();
-        }
-
-        private static void UpdateLevelsAndRewards()
-        {
-            Int16 max_level = 5;
-            Dictionary<String, RewardEntry> possible_rewards = Rewards.GetResouceRewards();
-            availableRewards = new List<String> { };
-            var zs = ZoneSystem.instance;
-            foreach (KeyValuePair<string, RewardEntry> entry in possible_rewards)
-            {
-                if(entry.Value.requiredBoss == "None") { 
-                    availableRewards.Add(entry.Key);
-                    continue;
-                }
-                if (!zs) 
-                {
-                    Jotunn.Logger.LogInfo("Zone system not available, skipping checks for global keys to set rewards options.");
-                    // We can only add items that do not require a global key check if there is no zone system.
-                    continue; 
-                }
-                if(entry.Value.requiredBoss == "Eikythr" && zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledEikthyr)) {
-                    availableRewards.Add(entry.Key);
-                    if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Killed Eikythr, enabling reward {entry.Key}."); }
-                    continue;
-                }
-                if (entry.Value.requiredBoss == "TheElder" && zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledElder))
-                {
-                    availableRewards.Add(entry.Key);
-                    if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Killed TheElder, enabling rewards {entry.Key}."); }
-                    continue;
-                }
-                if (entry.Value.requiredBoss == "BoneMass" && zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledBonemass))
-                {
-                    availableRewards.Add(entry.Key);
-                    if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Killed BoneMass, enabling rewards {entry.Key}."); }
-                    continue;
-                }
-                if (entry.Value.requiredBoss == "Moder" && zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledModer))
-                {
-                    availableRewards.Add(entry.Key);
-                    if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Killed Moder, enabling rewards {entry.Key}."); }
-                    continue;
-                }
-                if (entry.Value.requiredBoss == "Yagluth" && zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledYagluth))
-                {
-                    availableRewards.Add(entry.Key);
-                    if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Killed Yagluth, enabling rewards {entry.Key}."); }
-                    continue;
-                }
-                if (entry.Value.requiredBoss == "TheQueen" && zs.GetGlobalKey("defeated_queen"))
-                {
-                    availableRewards.Add(entry.Key);
-                    if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Killed TheQueen, enabling rewards {entry.Key}."); }
-                    continue;
-                }
-            }
-
-            if (zs) // If the zonesystem does not exist, we can't check global keys. So skip it. This list will be updated on the UI open anyways.
-            {
-                if (zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledEikthyr)) { max_level += 5; }
-                if (zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledElder)) { max_level += 5; }
-                if (zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledBonemass)) { max_level += 5; }
-                if (zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledModer)) { max_level += 5; }
-                if (zs.GetGlobalKey(Jotunn.Utils.GameConstants.GlobalKey.KilledYagluth)) { max_level += 5; }
-                if (zs.GetGlobalKey("defeated_queen")) { max_level += 5; }
-            } else
-            {
-                Jotunn.Logger.LogInfo("Zone system not available, skipping checks for global keys to increase max shrine levels.");
-            }
-            // If you have killed all of the tracked bosses, set the max possible level to 50.
-            // if (max_level == 35) { max_level = 50; }
-
-            // If the max challenge level is bigger than the level we determined we lower it to that
-            if (max_level > (short)VFConfig.MaxChallengeLevel.Value) { max_level = (short)VFConfig.MaxChallengeLevel.Value; }
-
-            // Empty out the current levels if it exists so we don't get duplicates
-            currentLevels = new List<String> { };
-            // Toss in all of the available levels
-            String biome_or_boss = "";
-            for (int i = 1; i <= max_level; i++)
-            {
-                if (i > 0 && i < 6) { biome_or_boss = Localization.instance.Localize("$shrine_menu_meadow"); }
-                if (i > 5 && i < 11) { biome_or_boss = Localization.instance.Localize("$shrine_menu_forest"); }
-                if (i > 10 && i < 16) { biome_or_boss = Localization.instance.Localize("$shrine_menu_swamp"); }
-                if (i > 15 && i < 21) { biome_or_boss = Localization.instance.Localize("$shrine_menu_mountain"); }
-                if (i > 20 && i < 26) { biome_or_boss = Localization.instance.Localize("$shrine_menu_plains"); }
-                if (i > 25 && i < 31) { biome_or_boss = Localization.instance.Localize("$shrine_menu_mistland"); }
-                currentLevels.Add($"{i} - {biome_or_boss}");
-            }
-            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo("Levels and rewards updated."); }
-        }
-
         private void StartChallenge()
         {
             HideUI();
@@ -224,7 +113,7 @@ namespace ValheimFortress.Challenge
             if (VFConfig.EnableSiegeModifer.Value) { siege_mode = siegeModeToggle.GetComponent<Toggle>().isOn; }
             if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Shrine challenge. Selected reward: {selected_reward}, selected level: {selected_level}"); }
             // Start the coroutine that sends the warning text
-            PreparePhase(selected_level, boss_mode, Shrine.gameObject);
+            UserInterfaceData.PreparePhase(selected_level, boss_mode, Shrine.gameObject);
             Shrine.SetLevel(selected_level);
             Shrine.SetReward(selected_reward);
             if (hard_mode) { Shrine.SetHardMode(); }
@@ -234,58 +123,9 @@ namespace ValheimFortress.Challenge
             Shrine.SetStartChallenge();
         }
 
-        private static void PreparePhase(Int16 selected_level, bool boss_mode, GameObject shrine)
-        {
-            String challenge_warning = "This might hurt.";
-            if (selected_level < 6) {
-                challenge_warning = Localization.instance.Localize("$shrine_warning_meadows");
-                if (boss_mode) { challenge_warning = Localization.instance.Localize("$shrine_warning_meadows_boss"); }
-            }
-            if (selected_level > 5 && selected_level < 11) { 
-                challenge_warning = Localization.instance.Localize("$shrine_warning_forest");
-                if (boss_mode) { challenge_warning = Localization.instance.Localize("$shrine_warning_forest_boss"); }
-            }
-            if (selected_level > 10 && selected_level < 16) { 
-                challenge_warning = Localization.instance.Localize("$shrine_warning_swamp");
-                if (boss_mode) { challenge_warning = Localization.instance.Localize("$shrine_warning_swamp_boss"); }
-            }
-            if (selected_level > 15 && selected_level < 21) { 
-                challenge_warning = Localization.instance.Localize("$shrine_warning_mountain");
-                if (boss_mode) { challenge_warning = Localization.instance.Localize("$shrine_warning_mountain_boss"); }
-            }
-            if (selected_level > 20 && selected_level < 26) { 
-                challenge_warning = Localization.instance.Localize("$shrine_warning_plains");
-                if (boss_mode) { challenge_warning = Localization.instance.Localize("$shrine_warning_plains_boss"); }
-            }
-            if (selected_level > 25 && selected_level < 31) {
-                challenge_warning = Localization.instance.Localize("$shrine_warning_mistlands");
-                if (boss_mode) { challenge_warning = Localization.instance.Localize("$shrine_warning_mistlands_boss"); }
-            }
-            List<Player> nearby_players = new List<Player> { };
-            Player.GetPlayersInRange(shrine.transform.position, VFConfig.ShrineAnnouncementRange.Value, nearby_players);
-            foreach(Player localplayer in nearby_players)
-            {
-                localplayer.Message(MessageHud.MessageType.Center, challenge_warning);
-            }
-            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo("Activated Shrine portal & sent warning message"); }
-        }
-
-        public static void PhasePausePhrase(GameObject shrine)
-        {
-            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Picking and sending phase waiting text from {shrine_phase_warnings.Count} phrases."); }
-            string selected_message = shrine_phase_warnings[UnityEngine.Random.Range(0, (shrine_phase_warnings.Count -1))];
-            List<Player> nearby_players = new List<Player> { };
-            Player.GetPlayersInRange(shrine.transform.position, VFConfig.ShrineAnnouncementRange.Value, nearby_players);
-            foreach (Player localplayer in nearby_players)
-            {
-                localplayer.Message(MessageHud.MessageType.Center, selected_message);
-            }
-        }
-
-
         public void DisplayUI()
         {
-            CreateStaticUIObjects(true);
+            CreateStaticUIObjects();
             CreateChallengeUI();
             ChallengePanel.SetActive(true);
             GUIManager.BlockInput(true);
@@ -300,11 +140,12 @@ namespace ValheimFortress.Challenge
             Shrine.SetShrineUIStatus(false);
         }
 
-        public void CreateStaticUIObjects(bool should_update_ui)
+        public void CreateStaticUIObjects()
         {
             // This was supposed to allow this is not need to be recreated much/at all
-            // However, there are a few edge cases that cause this to b
-            if (should_update_ui == false) { return; }
+            // However, there are a few edge cases that cause this to be null
+            // if (ChallengePanel != null) { return; }
+
             if (GUIManager.Instance == null)
             {
                 Jotunn.Logger.LogError("GUIManager instance is null");
@@ -436,7 +277,7 @@ namespace ValheimFortress.Challenge
                 addContentSizeFitter: false);
 
             // Create the start button object
-            GameObject startButtonObj = GUIManager.Instance.CreateButton(
+            startChallengeButtonGO = GUIManager.Instance.CreateButton(
                 text: Localization.instance.Localize("$shrine_confirm"),
                 parent: ChallengePanel.transform,
                 anchorMin: new Vector2(0.5f, 0.5f),
@@ -444,10 +285,10 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(0f, -224f),
                 width: 150f,
                 height: 60f);
-            startButtonObj.SetActive(true);
+            startChallengeButtonGO.SetActive(true);
 
             // Create the close button object
-            GameObject cancelButtonObj = GUIManager.Instance.CreateButton(
+            cancelButtonGO = GUIManager.Instance.CreateButton(
                 text: Localization.instance.Localize("$shrine_cancel"),
                 parent: ChallengePanel.transform,
                 anchorMin: new Vector2(0.5f, 0.5f),
@@ -455,22 +296,23 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(240f, 240f),
                 width: 40f,
                 height: 40f);
-            cancelButtonObj.SetActive(true);
-
-            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo("Adding UI Listeners"); }
+            cancelButtonGO.SetActive(true);
+            
+            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo("Shrine UI Created."); }
             // Add a listener to the button to close the panel again
-            Button cancelButton = cancelButtonObj.GetComponent<Button>();
+            Button cancelButton = cancelButtonGO.GetComponent<Button>();
             cancelButton.onClick.AddListener(HideUI);
             // Add a listener to the button to close the panel and trigger the challenge scripts
-            Button startButton = startButtonObj.GetComponent<Button>();
-            startButton.onClick.AddListener(StartChallenge);
-            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo("Shrine UI Created."); }
+            Button startChallengeButton = startChallengeButtonGO.GetComponent<Button>();
+            startChallengeButton.onClick.AddListener(StartChallenge);
+            if (VFConfig.EnableDebugMode.Value) { Jotunn.Logger.LogInfo("Adding UI Listeners"); }
+
         }
 
         public void CreateChallengeUI()
         {
             // Always want to update the rewards and challenge levels
-            UpdateLevelsAndRewards();
+            UserInterfaceData.UpdateLevelsAndRewards();
             // We specifically want to be able to completely rebuild the UI when this is called again to update the levels and rewards dynamically
 
             // Create the rewards selector dropdown
@@ -721,6 +563,7 @@ namespace ValheimFortress.Challenge
 
         public void createCancelUI()
         {
+            if (CancelPanel != null) { return; }
             // Create parent panel object
             CancelPanel = GUIManager.Instance.CreateWoodpanel(
                 parent: GUIManager.CustomGUIFront.transform,
@@ -730,8 +573,11 @@ namespace ValheimFortress.Challenge
                 width: 400,
                 height: 400,
                 draggable: true);
+            // This hides the panel immediately
+            CancelPanel.SetActive(false);
+
             // Create the close button object
-            GameObject cancelButtonObj = GUIManager.Instance.CreateButton(
+            GameObject cancelCancelButtonObj = GUIManager.Instance.CreateButton(
                 text: Localization.instance.Localize("$shrine_cancel"),
                 parent: CancelPanel.transform,
                 anchorMin: new Vector2(0.5f, 0.5f),
@@ -739,7 +585,6 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(175f, 175f),
                 width: 40f,
                 height: 40f);
-            cancelButtonObj.SetActive(true);
             // Create the title
             GUIManager.Instance.CreateText(
                 text: Localization.instance.Localize("$shrine_in_progress_header"),
@@ -793,7 +638,6 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(110f, -150f),
                 width: 100f,
                 height: 60f);
-            cancelChallengeObj.SetActive(true);
             GUIManager.Instance.CreateText(
                 text: Localization.instance.Localize("$shrine_cancel_challenge_desc"),
                 parent: CancelPanel.transform,
@@ -818,7 +662,6 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(110f, -75f),
                 width: 100f,
                 height: 40f);
-            cancelChallengeObj.SetActive(true);
             GUIManager.Instance.CreateText(
                 text: Localization.instance.Localize("$shrine_cleanup_portals_desc"),
                 parent: CancelPanel.transform,
@@ -843,7 +686,6 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(110f, 0f),
                 width: 100f,
                 height: 40f);
-            cancelChallengeObj.SetActive(true);
             GUIManager.Instance.CreateText(
                 text: Localization.instance.Localize("$shrine_enable_flares_desc"),
                 parent: CancelPanel.transform,
@@ -868,7 +710,6 @@ namespace ValheimFortress.Challenge
                 position: new Vector2(110f, 75f),
                 width: 100f,
                 height: 40f);
-            cancelChallengeObj.SetActive(true);
             GUIManager.Instance.CreateText(
                 text: Localization.instance.Localize("$shrine_teleport_creatures_desc"),
                 parent: CancelPanel.transform,
@@ -884,8 +725,8 @@ namespace ValheimFortress.Challenge
                 height: 40f,
             addContentSizeFitter: false);
 
-            Button cancelButton = cancelButtonObj.GetComponent<Button>();
-            cancelButton.onClick.AddListener(HideCancelUI);
+            Button cancelCancelButton = cancelCancelButtonObj.GetComponent<Button>();
+            cancelCancelButton.onClick.AddListener(HideCancelUI);
             Button cancelChallengeButton = cancelChallengeObj.GetComponent<Button>();
             cancelChallengeButton.onClick.AddListener(CancelChallengeButtonClick);
             Button cleanupPortalsButton = cleanupPortalsObj.GetComponent<Button>();
@@ -894,7 +735,6 @@ namespace ValheimFortress.Challenge
             enableFlaresButton.onClick.AddListener(AddCreatureFlares);
             Button teleportCreaturesButton = teleportCreaturesObj.GetComponent<Button>();
             teleportCreaturesButton.onClick.AddListener(TeleportCreatures);
-            
         }
 
         private void CleanupPortalsButtonClick()
@@ -913,6 +753,9 @@ namespace ValheimFortress.Challenge
 
         public void HideCancelUI()
         {
+            // Nothing to do if the UI doesn't exist yet
+            if (CancelPanel == null) { return; }
+
             CancelPanel.SetActive(false);
             GUIManager.BlockInput(false);
             Shrine.SetShrineUIStatus(false);
