+++ JotunnModStub/Localizations/Spanish.json
@@ -1,0 +1,113 @@
+﻿
+{
+ // Piezas de mármol blanco
+ "piece_vf_whitemarble_block": "Baldosa de mármol blanco",
+ // Tapetes
+ "piece_vf_carpet_circle_green": "Tapete circular de yute verde",
+ "piece_vf_carpet_circle_red": "Tapete circular de yute rojo",
+ "piece_vf_carpet_circle_yellow": "Tapete circular de yute amarillo",
+ // Segmentos de pared de cristal
+ "piece_vf_crystalwall_blue": "Pared de cristal azul",
+ "piece_vf_crystalwall_green": "Pared de cristal verde",
+ "piece_vf_crystalwall_red": "Pared de cristal rojo",
+ "piece_vf_crystalwall_yellow": "Pared de cristal amarillo",
+ // Ítems defensivos
+ "piece_stone_spikes": "Estacas de piedra",
+ "piece_vfturret": "Torreta automatizada",
+ "piece_vfturret_description": "Una torreta modificada que utiliza cantidades excesivas de magia para materializar munición, nunca necesita ser rearmando. También es un poco más inteligente al apuntar, pero no puede aceptar un filtro de objetivo.",
+ // Localizaciones para la interfaz de la Cámara del Desafío
+ "shrine_warning_meadows": "El prado ya no está en calma.",
+ "shrine_warning_meadows_boss": "Eikthyr está enfadado.",
+ "shrine_warning_forest": "El bosque se inquieta.",
+ "shrine_warning_forest_boss": "El bosque camina con El Anciano.",
+ "shrine_warning_swamp": "El pantano guarda a sus muertos.",
+ "shrine_warning_swamp_boss": "Bonemass se levanta del pantano.",
+ "shrine_warning_mountain": "La montaña resuena en aullidos.",
+ "shrine_warning_mountain_boss": "Moder desciende de la montaña.",
+ "shrine_warning_plains": "Las llanuras envían a sus bandas de guerra.",
+ "shrine_warning_plains_boss": "Yagluth se levanta en venganza.",
+ "shrine_warning_mistlands": "Las nieblas se apartan ante una horda que se arrastra.",
+ "shrine_warning_mistlands_boss": "La Reina acecha a su presa.",
+ "shrine_cancel_challenge": "Abandonar el Desafío",
+ "shrine_cancel_challenge_desc": "Cancela el desafío, mata a todas las criaturas restantes. Sin recompensa.",
+ "shrine_cleanup_portals": "Limpiar portales",
+ "shrine_cleanup_portals_desc": "Elimina los portales creados. No discrimina entre portales activos o antiguos.\nEliminar portales no interfiere con el desafío activo.",
+ "shrine_enable_flares": "Habilitar bengalas",
+ "shrine_enable_flares_desc": "Habilita bengalas para todas las criaturas restantes actualmente vivas.",
+ "shrine_teleport_creatures": "Teletransportar criaturas",
+ "shrine_teleport_creatures_desc": "Teletransporta las criaturas restantes si hay menos de 6, al santuario.",
+ "shrine_too_many_creautres_to_teleport": "Solo se pueden teletransportar menos de 6 criaturas. Restantes",
+ "shrine_cancel_enemies_remaining": "Enemigos restantes: ",
+ "shrine_in_progress_header": "Modo Desafío Activo",
+ // Localizaciones para los cortes de fase de la Cámara del Desafío
+ "$shrine_phase_warning": "Las criaturas se están reagrupando.",
+ "$shrine_phase_warning1": "Hay una pausa en los ataques.",
+ "$shrine_phase_warning2": "Hay un silencio extraño...",
+ "$shrine_phase_warning3": "Los portales están tranquilos por un momento...",
+ "$shrine_phase_warning4": "Un breve respiro de los ataques.",
+ "$shrine_phase_warning5": "¿Eso fue lo último de ellos?",
+ "$shrine_phase_warning6": "Los enemigos de Odín son numerosos...",
+ "$shrine_phase_warning7": "Parecen haberse retirado.",
+ "$shrine_phase_warning8": "Los portales aún brillan...",
+ "$shrine_phase_warning9": "Esta horda ha menguado...",
+ "$shrine_phase_warning10": "Hay una pequeña pausa.",
+ "$shrine_phase_warning11": "Los has rechazado, por ahora...",
+ "$shrine_phase_warning12": "Sus filas se están adelgazando.",
+ "$shrine_phase_warning13": "¿Cuántos quedan?",
+ "$shrine_phase_warning14": "¿Es ese el último?",
+ "$shrine_phase_warning15": "La victoria complace a Odín.",
+ "$shrine_phase_warning16": "Valhalla espera al digno...",
+ "$shrine_phase_warning17": "Los abandonados están enfurecidos...",
+ "$shrine_phase_warning18": "Has derrotado a este grupo.",
+ "$shrine_phase_warning19": "Odín observa tu progreso...",
+ "$shrine_challenge_complete": "¡Desafío completado!",
+ // Pieza de la Cámara del Desafío
+ "piece_shrine_of_challenge": "Santuario del desafío",
+ // Menú selector de nivel de la Cámara del Desafío
+ "shrine_menu_meadow": "Prado",
+ "shrine_menu_forest": "Bosque Negro",
+ "shrine_menu_swamp": "Pantano",
+ "shrine_menu_mountain": "Montañas",
+ "shrine_menu_plains": "Llanuras",
+ "shrine_menu_mistland": "Tierras de Niebla",
+ "shrine_menu_ashland": "Tierras de Ceniza",
+ "shrine_menu_boss": "Jefe",
+ // Recompensas de la Cámara del Desafío
+ "reward_coin": "Moneda",
+ "reward_wood": "Madera",
+ "reward_copper": "Cobre",
+ "reward_tin": "Estaño",
+ "reward_iron": "Hierro",
+ "reward_silver": "Plata",
+ "reward_blackmetal": "Metal negro",
+ "reward_corewood": "Madera central",
+ "reward_finewood": "Madera fina",
+ "reward_guck": "Guck",
+ "reward_tar": "Alquitrán",
+ "reward_sap": "Savia",
+ "reward_softtissue": "Tejido blando",
+ // Texto estático de la Cámara del Desafío
+ "shrine_header": "Enfrenta a los enemigos de Odín",
+ "shrine_description": "Enfrenta un desafío abrumador a cambio de una recompensa deseada.\nSe pueden desbloquear más recompensas derrotando a los jefes del mundo.\nNiveles más altos dan recompensas más grandes.",
+ "shrine_warning": "Enfrentarás probabilidades abrumadoras.",
+ "shrine_reward_label": "Recompensa deseada",
+ "shrine_reward_estimate": "~",
+ "shrine_level_label": "Fuerza de la oleada",
+ "shrine_confirm": "¡A Valhalla!",
+ "shrine_cancel": "x",
+ "shrine_modifiers_label": "Modificadores",
+ "shrine_modifiers_rewards_label": "Recompensas aumentadas en",
+ "shrine_hard_mode_label": "Modo difícil",
+ "shrine_hard_mode_reward": "+50%",
+ "shrine_hard_mode_description": "Duplica el tamaño de cada oleada.",
+ "shrine_boss_mode": "Modo jefe",
+ "shrine_boss_mode_reward": "+25%",
+ "shrine_boss_mode_description": "Convierte cualquier nivel en una oleada de jefes.",
+ "shrine_siege_mode": "Modo asedio",
+ "shrine_siege_mode_reward": "+50%",
+ "shrine_siege_mode_description": "Más oleadas, más recompensas.",
+ // Tutorial del cuervo
+ "tutorial_shrine_of_challenge_text": "¡Escucha, guerrero! Puedes convocar hordas de los enemigos de Odín para que luchen aquí; él te recompensará generosamente.\n\nTen cuidado, ya que los enemigos de Odín son numerosos y podrías ser abrumado si los enfrentas solo. Sería mejor que prepares algunas trampas y fortificaciones.",
+ "tutorial_shrine_of_challenge_topic": "Hugin: Santuario del Desafío", 
+ "tutorial_shrine_of_challenge_label": "Hugin: Santuario del Desafío"
+}
