+++ JotunnModStub/Package/manifest.json
@@ -1,7 +1,7 @@
 ﻿{
   "name": "ValheimFortress",
   "description": "Build a base, summon a challenge, defeat it and reap the rewards!",
-  "version_number": "0.9.12",
+  "version_number": "0.9.14",
   "website_url": "https://discord.gg/Dmr9PQTy9m",
   "dependencies": [
     "ValheimModding-Jotunn-2.14.5",
