+++ JotunnModStub/ValheimFortress.cs
@@ -21,7 +21,7 @@ namespace ValheimFortress
     {
         public const string PluginGUID = "MidnightsFX.ValheimFortress";
         public const string PluginName = "ValheimFortress";
-        public const string PluginVersion = "0.9.12";
+        public const string PluginVersion = "0.9.14";
 
         AssetBundle EmbeddedResourceBundle;
         public VFConfig cfg;
